pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-code"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7,<9",
    "typer>=0.12.3",
    "rich==13.7.1",
    "python-dotenv==1.0.1",
    "langchain>=0.2.12",
    "langchain-community>=0.2.10",
    "langchain-anthropic>=0.2.4",
    "langchain-google-genai>=2.0.4",
    "pydantic>=2.7.4,<3",
    "ruamel.yaml==0.18.6",
    "pyfiglet==1.0.4",
    "mcp>=0.4.2",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-tavily>=0.1.0",
    "deepagents>=0.0.3",
    "genai-processors>=1.1.0",
    "langchain-openai>=0.3.0,<0.4.0",
    "langchain-ollama>=0.3.0,<0.4.0",
]

[project.scripts]
langcode = "langchain_code.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["langchain_code*"]

[tool.setuptools.package-data]
"langchain_code.config" = ["mcp.json"]
